Separate borrow analysis responsibilities
