+++ crates/hir/src/analysis/semantic/borrowck/canon.rs
@@ -9,10 +9,9 @@ use crate::{
     analysis::{
         HirAnalysisDb,
         semantic::{
-            FieldIndex, SBlockId, SConst, SLocalId, SemOrigin, SemanticCalleeRef, SemanticInstance,
-            VariantIndex,
+            SBlockId, SConst, SLocalId, SemOrigin, SemanticInstance, VariantIndex,
+            borrowck::ir::{NExpr, NSStmtKind},
             consts::{SemConstValue, scalar_int},
-            get_or_build_semantic_instance,
         },
         ty::{
             provider::{ProviderAddressSpace, ProviderKind},
@@ -24,21 +23,24 @@ use crate::{
 use num_traits::ToPrimitive;
 
 use super::{
-    analyses::BorrowSummaryMode,
-    check::{provisional_memory_summary_voucher, semantic_memory_summary_voucher},
     diagnostics::normalized_body_internal_diag,
     facts::NormalizedBodyFacts,
     ir::{
-        FreshAllocSite, MemoryAccessKind, MemorySummaryId, MemorySummaryItem, MemorySummaryTarget,
-        NBorrowRoot, NBorrowRootId, NEffectArg, NEffectArgValue, NExpr, NSPlace, NSPlaceRoot,
-        NSProjectionPath, NSStmt, NSStmtKind, NormalizedBindingLowering, NormalizedSemanticBody,
-        PointerAddressSpaces, SemanticBorrowDiagnostic,
+        FreshAllocSite, NBorrowRoot, NBorrowRootId, NSPlace, NSPlaceRoot, NSProjectionPath,
+        NormalizedBindingLowering, NormalizedSemanticBody, PointerAddressSpaces,
+        SemanticBorrowDiagnostic,
     },
     pointer::{
-        is_pointer_bearing_type, mem_array_carrier_suffix, path_with_projection, pointer_slots,
-        projection_result_ty, raw_pointer_pointee_suffix,
+        mem_array_carrier_suffix, path_with_projection, projection_result_ty,
+        raw_pointer_pointee_suffix,
     },
 };
+
+mod summary;
+mod transfer;
+
+pub(super) use summary::BorrowSummaryResolver;
+pub(super) use transfer::BorrowStateTransferCx;
 
 pub(super) fn address_spaces_for_borrow_root<'db>(
     db: &'db dyn HirAnalysisDb,
@@ -97,13 +99,6 @@ pub(super) enum BorrowRoot<'db> {
 pub(super) struct CanonPlace<'db> {
     pub(super) root: BorrowRoot<'db>,
     pub(super) proj: NSProjectionPath<'db>,
-}
-
-pub(super) struct ResolvedMemoryAccess<'db> {
-    pub(super) kind: MemoryAccessKind,
-    pub(super) targets: FxHashSet<CanonPlace<'db>>,
-    pub(super) authorized: FxHashSet<LoanId>,
-    pub(super) authorizer_targets: FxHashSet<CanonPlace<'db>>,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
@@ -525,14 +520,13 @@ impl JoinSemiLattice for State<'_> {
     }
 }
 
+#[derive(Clone, Copy)]
 pub(super) struct BorrowCanonCx<'a, 'db> {
     db: &'db dyn HirAnalysisDb,
     instance: SemanticInstance<'db>,
     body: &'a NormalizedSemanticBody<'db>,
     facts: &'a NormalizedBodyFacts,
     loans: &'a [Loan<'db>],
-    loan_for_local: &'a FxHashMap<SLocalId, LoanId>,
-    summary_mode: BorrowSummaryMode,
 }
 
 impl<'a, 'db> BorrowCanonCx<'a, 'db> {
@@ -542,8 +536,6 @@ impl<'a, 'db> BorrowCanonCx<'a, 'db> {
         body: &'a NormalizedSemanticBody<'db>,
         facts: &'a NormalizedBodyFacts,
         loans: &'a [Loan<'db>],
-        loan_for_local: &'a FxHashMap<SLocalId, LoanId>,
-        summary_mode: BorrowSummaryMode,
     ) -> Self {
         Self {
             db,
@@ -551,659 +543,6 @@ impl<'a, 'db> BorrowCanonCx<'a, 'db> {
             body,
             facts,
             loans,
-            loan_for_local,
-            summary_mode,
-        }
-    }
-
-    fn assign_pointer_targets_for_expr(
-        &self,
-        state: &mut State<'db>,
-        dst: SLocalId,
-        expr: &NExpr<'db>,
-        call_summary: Option<MemorySummaryId<'db>>,
-        origin: SemOrigin<'db>,
-    ) -> Result<(), SemanticBorrowDiagnostic<'db>> {
-        let Some(dst_ty) = self.body.local(dst).map(|local| local.ty) else {
-            return Ok(());
-        };
-        if dst_ty.as_borrow(self.db).is_some() || !is_pointer_bearing_type(self.db, dst_ty) {
-            return Ok(());
-        }
-        match expr {
-            NExpr::Use(src) => {
-                self.copy_pointer_targets_from_value(state, dst, src.local, dst_ty, origin)
-            }
-            NExpr::ReadPlace { place, .. } => {
-                self.copy_pointer_targets_from_place(state, dst, place, dst_ty, origin)
-            }
-            NExpr::AggregateMake { ty, fields } => {
-                self.assign_aggregate_pointer_targets(state, dst, *ty, fields, origin)
-            }
-            NExpr::ArrayRepeat { ty, value } => {
-                self.assign_array_repeat_pointer_targets(state, dst, *ty, *value, origin)
-            }
-            NExpr::EnumMake {
-                enum_ty,
-                variant,
-                fields,
-            } => self.assign_enum_pointer_targets(state, dst, *enum_ty, *variant, fields, origin),
-            NExpr::ExtractEnumField {
-                value,
-                variant,
-                field,
-            } => self
-                .copy_pointer_targets_from_enum_field(state, dst, *value, *variant, *field, origin),
-            NExpr::Cast { value, to } if is_pointer_bearing_type(self.db, *to) => {
-                let src_ty = self.body.local(value.local).map(|local| local.ty);
-                if src_ty.is_some_and(|ty| is_pointer_bearing_type(self.db, ty)) {
-                    self.copy_pointer_targets_from_value(state, dst, value.local, dst_ty, origin)
-                } else {
-                    // Non-pointer-to-pointer casts model raw address construction
-                    // such as integer-to-pointer. Address-of or borrow-to-pointer
-                    // lowering must preserve Local/Provider provenance before it
-                    // reaches this fallback.
-                    self.assign_unknown_pointer_targets(state, dst, dst_ty);
-                    Ok(())
-                }
-            }
-            NExpr::Call {
-                args, effect_args, ..
-            } => {
-                let summary = call_summary.expect("call expression must have a memory summary");
-                let dst_bases = self.canonicalize_value_base(state, dst);
-                let mut by_output: FxHashMap<NSProjectionPath<'db>, PointerTargets<'db>> =
-                    FxHashMap::default();
-                for item in summary.items(self.db) {
-                    let MemorySummaryItem::ReturnPointer { output, targets } = item else {
-                        continue;
-                    };
-                    let item_targets =
-                        self.resolve_summary_targets(state, args, effect_args, targets, origin)?;
-                    by_output
-                        .entry(output.clone())
-                        .and_modify(|existing| {
-                            existing.join_into(&item_targets);
-                        })
-                        .or_insert(item_targets);
-                }
-                for (output, targets) in by_output {
-                    for base in &dst_bases {
-                        let key =
-                            self.pointer_slot_place(base.root.clone(), base.proj.concat(&output));
-                        state.assign_pointer_targets(key, targets.clone());
-                    }
-                }
-                Ok(())
-            }
-            _ => {
-                self.assign_unknown_pointer_targets(state, dst, dst_ty);
-                Ok(())
-            }
-        }
-    }
-
-    fn resolve_summary_targets(
-        &self,
-        state: &State<'db>,
-        args: &[super::ir::NOperand],
-        effect_args: &[NEffectArg<'db>],
-        summary: &[MemorySummaryTarget<'db>],
-        origin: SemOrigin<'db>,
-    ) -> Result<PointerTargets<'db>, SemanticBorrowDiagnostic<'db>> {
-        let mut out = PointerTargets::default();
-        for target in summary {
-            out.join_into(
-                &self
-                    .resolve_summary_target(state, args, effect_args, target, origin)?
-                    .0,
-            );
-        }
-        Ok(out)
-    }
-
-    fn resolve_summary_target(
-        &self,
-        state: &State<'db>,
-        args: &[super::ir::NOperand],
-        effect_args: &[NEffectArg<'db>],
-        target: &MemorySummaryTarget<'db>,
-        origin: SemOrigin<'db>,
-    ) -> Result<(PointerTargets<'db>, FxHashSet<LoanId>), SemanticBorrowDiagnostic<'db>> {
-        Ok(match target {
-            MemorySummaryTarget::Input { input, proj } => {
-                let super::ir::BorrowInputRef::Param(idx) = input;
-                let Some(arg) = args.get(*idx as usize) else {
-                    return Err(self.internal_diag(
-                        origin,
-                        format!("memory summary refers to missing parameter {idx}"),
-                    ));
-                };
-                (
-                    PointerTargets::known(
-                        self.canonicalize_value_path(state, arg.local, proj, origin)?,
-                    ),
-                    self.loans_for_value(state, arg.local),
-                )
-            }
-            MemorySummaryTarget::Effect { binding_idx, proj } => {
-                let Some(arg) = effect_args
-                    .iter()
-                    .find(|arg| arg.binding_idx == *binding_idx)
-                else {
-                    return Err(self.internal_diag(
-                        origin,
-                        format!("memory summary refers to missing effect argument {binding_idx}"),
-                    ));
-                };
-                match &arg.arg {
-                    NEffectArgValue::Place(place) => {
-                        let bases = self.canonicalize_place(state, place, origin)?;
-                        (
-                            PointerTargets::known(
-                                self.canonicalize_path_from_places(&bases, proj, state, origin)?,
-                            ),
-                            self.loans_for_place(state, place),
-                        )
-                    }
-                    NEffectArgValue::Value(value) => (
-                        PointerTargets::known(self.canonicalize_value_path(
-                            state,
-                            value.local,
-                            proj,
-                            origin,
-                        )?),
-                        self.loans_for_value(state, value.local),
-                    ),
-                }
-            }
-            MemorySummaryTarget::FreshAllocation {
-                site,
-                address_space,
-            } => (
-                PointerTargets::known(FxHashSet::from_iter([CanonPlace {
-                    root: BorrowRoot::FreshAllocation {
-                        site: FreshAllocSite::Call {
-                            call: origin,
-                            callee: match site {
-                                FreshAllocSite::Direct(origin)
-                                | FreshAllocSite::Call { call: origin, .. } => *origin,
-                            },
-                        },
-                        address_space: *address_space,
-                    },
-                    proj: NSProjectionPath::default(),
-                }])),
-                FxHashSet::default(),
-            ),
-            MemorySummaryTarget::Unknown { address_spaces } => (
-                PointerTargets::unknown(*address_spaces),
-                FxHashSet::default(),
-            ),
-        })
-    }
-
-    fn callee_memory_summary(
-        &self,
-        callee: &SemanticCalleeRef<'db>,
-    ) -> Result<MemorySummaryId<'db>, SemanticBorrowDiagnostic<'db>> {
-        let callee_instance = get_or_build_semantic_instance(self.db, callee.key);
-        match self.summary_mode.for_callee(self.db, callee_instance) {
-            BorrowSummaryMode::Final => semantic_memory_summary_voucher(self.db, callee_instance),
-            BorrowSummaryMode::Provisional => {
-                provisional_memory_summary_voucher(self.db, callee_instance)
-            }
-        }
-    }
-
-    pub(super) fn resolve_call_memory_accesses(
-        &self,
-        state: &State<'db>,
-        callee: &SemanticCalleeRef<'db>,
-        args: &[super::ir::NOperand],
-        effect_args: &[NEffectArg<'db>],
-        origin: SemOrigin<'db>,
-    ) -> Result<Vec<ResolvedMemoryAccess<'db>>, SemanticBorrowDiagnostic<'db>> {
-        let summary = self.callee_memory_summary(callee)?;
-        let mut out = Vec::new();
-        for item in summary.items(self.db) {
-            let MemorySummaryItem::Access {
-                target,
-                kind,
-                authorizers,
-            } = item
-            else {
-                continue;
-            };
-            let (targets, loans) =
-                self.resolve_summary_target(state, args, effect_args, target, origin)?;
-            // Direct accesses are authorized only by mutable loans. Explicit
-            // authorizers identify the receiver/capability through which an
-            // otherwise unknown effect occurs, so their immutable loans must
-            // also be excluded from self-conflict checks.
-            let mut authorized = self.mut_loans(loans);
-            let mut authorizer_targets = FxHashSet::default();
-            for authorizer in authorizers {
-                let (targets, loans) =
-                    self.resolve_summary_target(state, args, effect_args, authorizer, origin)?;
-                authorized.extend(loans);
-                authorizer_targets.extend(targets.places());
-            }
-            out.push(ResolvedMemoryAccess {
-                kind: *kind,
-                targets: targets.places(),
-                authorized,
-                authorizer_targets,
-            });
-        }
-        Ok(out)
-    }
-
-    fn apply_call_memory_summary(
-        &self,
-        state: &mut State<'db>,
-        args: &[super::ir::NOperand],
-        effect_args: &[NEffectArg<'db>],
-        summary: MemorySummaryId<'db>,
-        origin: SemOrigin<'db>,
-    ) -> Result<bool, SemanticBorrowDiagnostic<'db>> {
-        let mut resolved = Vec::new();
-        for item in summary.items(self.db) {
-            let MemorySummaryItem::StorePointer {
-                target,
-                targets,
-                weak,
-            } = item
-            else {
-                continue;
-            };
-            let written = self
-                .resolve_summary_target(state, args, effect_args, target, origin)?
-                .0
-                .places();
-            let targets =
-                self.resolve_summary_targets(state, args, effect_args, targets, origin)?;
-            resolved.push((written, targets, *weak));
-        }
-        for (written, targets, weak) in resolved {
-            let weak = weak || self.write_precision(&written) == WritePrecision::Weak;
-            for place in written {
-                state.update_pointer_targets(
-                    self.pointer_slot_place(place.root, place.proj),
-                    targets.clone(),
-                    weak,
-                );
-            }
-        }
-        Ok(summary.may_return(self.db))
-    }
-
-    fn clear_pointer_targets_for_local(&self, state: &mut State<'db>, local: SLocalId) {
-        let Some(local_data) = self.body.local(local) else {
-            return;
-        };
-        for base in self.structural_local_places(local) {
-            self.clear_physical_pointer_slots(state, &base, local_data.ty);
-        }
-    }
-
-    fn copy_pointer_targets_from_value(
-        &self,
-        state: &mut State<'db>,
-        dst: SLocalId,
-        src: SLocalId,
-        dst_ty: TyId<'db>,
-        origin: SemOrigin<'db>,
-    ) -> Result<(), SemanticBorrowDiagnostic<'db>> {
-        let dst_bases = self.canonicalize_value_base(state, dst);
-        for slot in pointer_slots(self.db, dst_ty) {
-            let targets = self.pointer_targets_for_value_path(state, src, &slot.path, origin)?;
-            for base in &dst_bases {
-                state.assign_pointer_targets(
-                    self.pointer_slot_place(base.root.clone(), base.proj.concat(&slot.path)),
-                    targets.clone(),
-                );
-            }
-        }
-        Ok(())
-    }
-
-    fn copy_pointer_targets_from_place(
-        &self,
-        state: &mut State<'db>,
-        dst: SLocalId,
-        src: &NSPlace<'db>,
-        dst_ty: TyId<'db>,
-        origin: SemOrigin<'db>,
-    ) -> Result<(), SemanticBorrowDiagnostic<'db>> {
-        let dst_bases = self.canonicalize_value_base(state, dst);
-        let src_bases = self.canonicalize_place(state, src, origin)?;
-        for slot in pointer_slots(self.db, dst_ty) {
-            let keys = self.canonicalize_path_from_places(&src_bases, &slot.path, state, origin)?;
-            let targets = self.pointer_targets_for_keys(state, keys, origin)?;
-            for base in &dst_bases {
-                state.assign_pointer_targets(
-                    self.pointer_slot_place(base.root.clone(), base.proj.concat(&slot.path)),
-                    targets.clone(),
-                );
-            }
-        }
-        Ok(())
-    }
-
-    fn assign_array_repeat_pointer_targets(
-        &self,
-        state: &mut State<'db>,
-        dst: SLocalId,
-        ty: TyId<'db>,
-        value: super::ir::NOperand,
-        origin: SemOrigin<'db>,
-    ) -> Result<(), SemanticBorrowDiagnostic<'db>> {
-        let dst_bases = self.canonicalize_value_base(state, dst);
-        for slot in pointer_slots(self.db, ty) {
-            let Some(suffix) = split_array_slot(&slot.path, false) else {
-                continue;
-            };
-            let targets =
-                self.pointer_targets_for_value_path(state, value.local, &suffix, origin)?;
-            for base in &dst_bases {
-                state.assign_pointer_targets(
-                    self.pointer_slot_place(base.root.clone(), base.proj.concat(&slot.path)),
-                    targets.clone(),
-                );
-            }
-        }
-        Ok(())
-    }
-
-    fn assign_aggregate_pointer_targets(
-        &self,
-        state: &mut State<'db>,
-        dst: SLocalId,
-        ty: TyId<'db>,
-        fields: &[super::ir::NOperand],
-        origin: SemOrigin<'db>,
-    ) -> Result<(), SemanticBorrowDiagnostic<'db>> {
-        let dst_bases = self.canonicalize_value_base(state, dst);
-        for slot in pointer_slots(self.db, ty) {
-            let targets = if let Some((field_idx, suffix)) = split_aggregate_slot(&slot.path) {
-                let Some(field) = fields.get(field_idx) else {
-                    continue;
-                };
-                self.pointer_targets_for_value_path(state, field.local, &suffix, origin)?
-            } else if let Some(suffix) = split_array_slot(&slot.path, true) {
-                let mut targets = PointerTargets::default();
-                for field in fields {
-                    let field_targets =
-                        self.pointer_targets_for_value_path(state, field.local, &suffix, origin)?;
-                    targets.join_into(&field_targets);
-                }
-                targets
-            } else {
-                continue;
-            };
-            for base in &dst_bases {
-                state.assign_pointer_targets(
-                    self.pointer_slot_place(base.root.clone(), base.proj.concat(&slot.path)),
-                    targets.clone(),
-                );
-            }
-        }
-        Ok(())
-    }
-
-    fn assign_enum_pointer_targets(
-        &self,
-        state: &mut State<'db>,
-        dst: SLocalId,
-        enum_ty: TyId<'db>,
-        variant: VariantIndex,
-        fields: &[super::ir::NOperand],
-        origin: SemOrigin<'db>,
-    ) -> Result<(), SemanticBorrowDiagnostic<'db>> {
-        let dst_bases = self.canonicalize_value_base(state, dst);
-        for slot in pointer_slots(self.db, enum_ty) {
-            let Some((field_idx, suffix)) = split_enum_slot(&slot.path, variant) else {
-                continue;
-            };
-            let Some(field) = fields.get(field_idx) else {
-                continue;
-            };
-            let targets =
-                self.pointer_targets_for_value_path(state, field.local, &suffix, origin)?;
-            for base in &dst_bases {
-                state.assign_pointer_targets(
-                    self.pointer_slot_place(base.root.clone(), base.proj.concat(&slot.path)),
-                    targets.clone(),
-                );
-            }
-        }
-        Ok(())
-    }
-
-    fn copy_pointer_targets_from_enum_field(
-        &self,
-        state: &mut State<'db>,
-        dst: SLocalId,
-        value: super::ir::NOperand,
-        variant: VariantIndex,
-        field: FieldIndex,
-        origin: SemOrigin<'db>,
-    ) -> Result<(), SemanticBorrowDiagnostic<'db>> {
-        let Some(enum_ty) = self.body.local(value.local).map(|local| local.ty) else {
-            return Ok(());
-        };
-        let Some(dst_ty) = self.body.local(dst).map(|local| local.ty) else {
-            return Ok(());
-        };
-        let field_path = NSProjectionPath::from_projection(Projection::VariantField {
-            enum_ty,
-            variant,
-            field_idx: field.0 as usize,
-        });
-        let dst_bases = self.canonicalize_value_base(state, dst);
-        for slot in pointer_slots(self.db, dst_ty) {
-            let path = field_path.concat(&slot.path);
-            let targets = self.pointer_targets_for_value_path(state, value.local, &path, origin)?;
-            for base in &dst_bases {
-                state.assign_pointer_targets(
-                    self.pointer_slot_place(base.root.clone(), base.proj.concat(&slot.path)),
-                    targets.clone(),
-                );
-            }
-        }
-        Ok(())
-    }
-
-    fn assign_unknown_pointer_targets(
-        &self,
-        state: &mut State<'db>,
-        dst: SLocalId,
-        dst_ty: TyId<'db>,
-    ) {
-        let dst_bases = self.canonicalize_value_base(state, dst);
-        for slot in pointer_slots(self.db, dst_ty) {
-            for base in &dst_bases {
-                state.assign_pointer_targets(
-                    self.pointer_slot_place(base.root.clone(), base.proj.concat(&slot.path)),
-                    PointerTargets::unknown(PointerAddressSpaces::one(
-                        ProviderAddressSpace::Memory,
-                    )),
-                );
-            }
-        }
-    }
-
-    pub(super) fn seed_param_pointer_targets(
-        &self,
-        state: &mut State<'db>,
-        local: SLocalId,
-        param_idx: u32,
-    ) {
-        let Some(local_data) = self.body.local(local) else {
-            return;
-        };
-        let base = CanonPlace {
-            root: BorrowRoot::Param(param_idx),
-            proj: NSProjectionPath::default(),
-        };
-        self.seed_structural_pointer_targets(state, base, local_data.ty);
-    }
-
-    fn seed_structural_pointer_targets(
-        &self,
-        state: &mut State<'db>,
-        base: CanonPlace<'db>,
-        ty: TyId<'db>,
-    ) {
-        for slot in pointer_slots(self.db, ty) {
-            let key = self.pointer_slot_place(base.root.clone(), base.proj.concat(&slot.path));
-            let target = CanonPlace {
-                root: base.root.clone(),
-                proj: base.proj.concat(&slot.path).concat(&slot.target_suffix),
-            };
-            state
-                .assign_pointer_targets(key, PointerTargets::known(FxHashSet::from_iter([target])));
-        }
-    }
-
-    pub(super) fn apply_stmt_state(
-        &self,
-        state: &mut State<'db>,
-        stmt: &NSStmt<'db>,
-    ) -> Result<(), SemanticBorrowDiagnostic<'db>> {
-        match &stmt.kind {
-            NSStmtKind::Assign { dst, expr } => {
-                let call_summary = if let NExpr::Call {
-                    callee,
-                    args,
-                    effect_args,
-                    ..
-                } = expr
-                {
-                    let summary = self.callee_memory_summary(callee)?;
-                    if !self.apply_call_memory_summary(
-                        state,
-                        args,
-                        effect_args,
-                        summary,
-                        stmt.origin,
-                    )? {
-                        state.mark_unreachable();
-                        return Ok(());
-                    }
-                    Some(summary)
-                } else {
-                    None
-                };
-                if self
-                    .body
-                    .local(*dst)
-                    .is_some_and(|local| local.ty.as_borrow(self.db).is_none())
-                {
-                    self.clear_pointer_targets_for_local(state, *dst);
-                }
-                let loans = match expr {
-                    NExpr::Use(src) => {
-                        let loans = state.loans_in(src.local);
-                        if loans.is_empty() {
-                            self.loan_for_local
-                                .get(dst)
-                                .copied()
-                                .map(|loan| FxHashSet::from_iter([loan]))
-                                .unwrap_or_default()
-                        } else {
-                            loans
-                        }
-                    }
-                    NExpr::Borrow { .. } | NExpr::Call { .. } => self
-                        .loan_for_local
-                        .get(dst)
-                        .copied()
-                        .map(|loan| FxHashSet::from_iter([loan]))
-                        .unwrap_or_default(),
-                    _ => FxHashSet::default(),
-                };
-                state.assign_loans(*dst, loans);
-                self.assign_pointer_targets_for_expr(state, *dst, expr, call_summary, stmt.origin)
-            }
-            NSStmtKind::Store { dst, src } => {
-                let Some(src_ty) = self.body.local(src.local).map(|local| local.ty) else {
-                    return Ok(());
-                };
-                let written = self.canonicalize_place(state, dst, stmt.origin)?;
-                let precision = self.write_precision(&written);
-                if !is_pointer_bearing_type(self.db, src_ty) {
-                    self.invalidate_scalar_store_pointer_targets(state, &written, precision);
-                    return Ok(());
-                }
-                if precision == WritePrecision::Strong {
-                    for place in &written {
-                        self.clear_physical_pointer_slots(state, place, src_ty);
-                    }
-                }
-                for slot in pointer_slots(self.db, src_ty) {
-                    let targets = self.pointer_targets_for_value_path(
-                        state,
-                        src.local,
-                        &slot.path,
-                        stmt.origin,
-                    )?;
-                    for base in &written {
-                        let key = self
-                            .pointer_slot_place(base.root.clone(), base.proj.concat(&slot.path));
-                        state.update_pointer_targets(
-                            key,
-                            targets.clone(),
-                            precision == WritePrecision::Weak,
-                        );
-                    }
-                }
-                Ok(())
-            }
-        }
-    }
-
-    fn invalidate_scalar_store_pointer_targets(
-        &self,
-        state: &mut State<'db>,
-        written: &FxHashSet<CanonPlace<'db>>,
-        precision: WritePrecision,
-    ) {
-        for place in written {
-            let key = self.pointer_slot_place(place.root.clone(), place.proj.clone());
-            state.invalidate_pointer_targets(
-                key,
-                PointerTargets::unknown(PointerAddressSpaces::one(ProviderAddressSpace::Memory)),
-                precision == WritePrecision::Weak,
-            );
-        }
-    }
-
-    fn write_precision(&self, written: &FxHashSet<CanonPlace<'db>>) -> WritePrecision {
-        if written.len() == 1
-            && written.iter().all(|place| {
-                self.pointer_slot_place(place.root.clone(), place.proj.clone())
-                    .is_precise_slot()
-            })
-        {
-            WritePrecision::Strong
-        } else {
-            WritePrecision::Weak
-        }
-    }
-
-    fn clear_physical_pointer_slots(
-        &self,
-        state: &mut State<'db>,
-        base: &CanonPlace<'db>,
-        ty: TyId<'db>,
-    ) {
-        for slot in pointer_slots(self.db, ty) {
-            let key = self.pointer_slot_place(base.root.clone(), base.proj.concat(&slot.path));
-            if !key.may_name_multiple_slots() {
-                state.clear_pointer_target(&key);
-            }
         }
     }
 
@@ -1944,58 +1283,4 @@ fn address_spaces_for_root_conservative<'db>(root: &BorrowRoot<'db>) -> PointerA
                 }
             }),
     }
-}
-
-fn split_aggregate_slot<'db>(
-    path: &NSProjectionPath<'db>,
-) -> Option<(usize, NSProjectionPath<'db>)> {
-    let mut iter = path.iter();
-    let field_idx = match iter.next()? {
-        Projection::Field(field_idx) => *field_idx,
-        Projection::Index(IndexSource::Constant(idx)) => *idx,
-        _ => return None,
-    };
-    let mut suffix = NSProjectionPath::default();
-    for projection in iter {
-        suffix.push(projection.clone());
-    }
-    Some((field_idx, suffix))
-}
-
-fn split_array_slot<'db>(
-    path: &NSProjectionPath<'db>,
-    wildcard_only: bool,
-) -> Option<NSProjectionPath<'db>> {
-    let mut iter = path.iter();
-    let is_array_slot = match iter.next()? {
-        Projection::Index(IndexSource::Any) => true,
-        Projection::Index(_) => !wildcard_only,
-        _ => false,
-    };
-    if !is_array_slot {
-        return None;
-    }
-    let mut suffix = NSProjectionPath::default();
-    for projection in iter {
-        suffix.push(projection.clone());
-    }
-    Some(suffix)
-}
-
-fn split_enum_slot<'db>(
-    path: &NSProjectionPath<'db>,
-    expected_variant: VariantIndex,
-) -> Option<(usize, NSProjectionPath<'db>)> {
-    let mut iter = path.iter();
-    let field_idx = match iter.next()? {
-        Projection::VariantField {
-            variant, field_idx, ..
-        } if *variant == expected_variant => *field_idx,
-        _ => return None,
-    };
-    let mut suffix = NSProjectionPath::default();
-    for projection in iter {
-        suffix.push(projection.clone());
-    }
-    Some((field_idx, suffix))
 }
